requirements.txt
dj-rest-auth==2.2.6
Django==4.0.8
django-allauth==0.52.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
Faker==16.4.0
gunicorn==20.1.0
Markdown==3.4.1
psycopg2